Python: Fixup CWE-078 tests.
