--- python/ql/test/experimental/query-tests/Security-new-dataflow/CWE-078/CommandInjection.qlref
@@ -1,1 +1,0 @@
-experimental/Security-new-dataflow/CWE-078/CommandInjection.ql
