+++ python/ql/test/query-tests/Security/CWE-078-py2/CommandInjection.qlref
@@ -1,0 +1,1 @@
+Security/CWE-078/CommandInjection.ql
