+++ python/ql/test/query-tests/Security/CWE-078-py2/options
